return crop weight to double

+++ src/main/java/org/example/models/RootCrop.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 
 public class RootCrop implements Serializable {
     private String type;
-    private int weight;
+    private double weight;
     private String color;
 
     RootCrop(RootCropBuilder rootCropBuilder) {
@@ -17,9 +17,11 @@ public class RootCrop implements Serializable {
     public String getType() {
         return type;
     }
-    public Integer getWeight() {
+
+    public double getWeight() {
         return weight;
     }
+
     public String getColor() {
         return color;
     }
@@ -39,22 +41,24 @@ public class RootCrop implements Serializable {
 
     @Override
     public String toString() {
-        return String.format("Корнеплод: Тип: %s, Вес: %d, Цвет: %s", getType(), getWeight(), getColor());
+        return String.format("Корнеплод: Тип: %s, Вес: %f, Цвет: %s", getType(), getWeight(), getColor());
     }
 
     public static class RootCropBuilder {
         private String type;
-        private int weight;
+        private double weight;
         private String color;
 
-        public RootCropBuilder setType(String type){
+        public RootCropBuilder setType(String type) {
             this.type = type;
             return this;
         }
-        public RootCropBuilder setWeight(int weight){
+
+        public RootCropBuilder setWeight(double weight) {
             this.weight = weight;
             return this;
         }
+
         public RootCropBuilder setColor(String color) {
             this.color = color;
             return this;
